requirements.txt
python-dotenv
python-telegram-bot
paho-mqtt
nest_asyncio
